backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pyjwt==2.10.1
bcrypt==5.0.0
python-multipart==0.0.19
httpx==0.28.1
openai==1.58.1
oss2==2.19.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
email-validator==2.3.0